requirements.txt
Flask-SQLAlchemy==3.1.1
Flask-Cors==3.0.10
Werkzeug==2.3.0